+++ CHANGELOG.md
@@ -12,9 +12,13 @@ Added:
 - HN62402 (128KBx16/256x8) support for fire-40-a.  Uses 512KB on flash.
 - Support for prototypes fire-28-c and fire-32-b.
 - SST39SF040 support (fire-32-b only).  Uses 512KB on flash.
+- Firmware decoding support from
 
 Fixed:
 - 23C1010 support - there were failures when creating firmware with 23C1010.
+
+Retired:
+- `sdrr-info` (superceded by the CLI `onerom firmware inspect` command)
 
 ## v0.6.12 - 2026-05-26